[OR-198] Divide tokamak-fee into register-fee-token and fee-token-sendtx
