+++ examples/fee-token-sendtx/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tokamak-fee",
+  "name": "fee-token-sendtx",
   "version": "1.0.0",
   "scripts": {
     "clean": "rm -rf node_modules && rm yarn.lock",
