+++ examples/reigster-fee-token/.env.example
@@ -1,0 +1,4 @@
+# hardhat-remote
+L1_NODE_WEB3_URL=http://dev.testnet.optimism.tokamak.network:8545
+L2_NODE_WEB3_URL=http://localhost:8545
+PRIV_KEY=0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80
