+++ examples/reigster-fee-token/utils.js
@@ -1,0 +1,5 @@
+
+
+function sleep(ms) {
+  return new Promise((r) => setTimeout(r, ms))
+}
